post controller  수정 삭제 변경

+++ src/main/java/com/project/hamsterd/service/PostService.java
@@ -22,7 +22,7 @@ public class PostService {
     }
 
     public Post show(int postNo) {
-        log.info(postNo);
+        log.info("show postNo " + postNo);
         Post post = dao.findById(postNo).orElse(null);
         log.info(post);
 
@@ -56,19 +56,23 @@ public class PostService {
 
 
         Post create = dao.findById(post.getPostNo()).orElse(null);
-        post.setCreateTime( create.getCreateTime());
+        post.setCreateTime(create.getCreateTime());
 
         post.setUpdateTime(formattedDate);
-
-
-
-        return dao.save(post);
+        log.info("create: " + create);
+        if(create!=null){
+            return dao.save(post);
+        }
+        return null;
     }
 
     public Post delete(int postNo){
         Post data = dao.findById(postNo).orElse(null);
-        dao.delete(data);
-        return data;
+
+        if(data !=null) {
+            dao.delete(data);
+        }
+        return null;
     }
 
     public List<Post> findByMemberId(int id) {
